Сделан модуль с репозиторием (интерфейс можно использовать для мокирования данных)

+++ core/database/src/main/java/com/yasunov/database/entity/NoteEntity.kt
@@ -15,5 +15,5 @@ data class NoteEntity(
     @ColumnInfo(name = "uri_picture")
     val uriPicture: String,
     @ColumnInfo(name = "date_of_create")
-    val dateOfCreation: Int,
+    val dateOfCreation: Long,
 )
